Multimedia & Cleaning
1. Added database + wrapper to store 'multimedia'
a) Stores a thumbnail of the multimedia as a data URI
b) Stores the URL of the full media

2. Added a media loader, which will contain all functionality relating
to the media library - import to library, get library, remover from
library etc.
Currently can add an image or video (including generating of thumbnails)
No user interface at the moment. Need to figure out access of the local
filesystem using chromes FileSystem api

3. Currently a splitview displays a multimedia panel, which shows all
media loaded in the library. This is just for testing and we can move it
somewhere else when needed

4. Got rid of global variables in ccli stuff

5. Added a wrapper panel for the main display area, so that we can
change the content in this area.

6. Cleaned up presentationframespanel

7. Fixed ListPanel to use let instead of var in a loop where functions
are created